Remove tts and server state as class members of commands.
Small improvement of the base class Command as well as its
children. Removes, at construction time, the need to pass tts and server
state as class members, as this connects the life of those objects to
the commands for a long time. The commands are expected to live the
entire execution of the program, but this is not the case of tts or
server state. In case we load a different tts engine, or a different
voice, we may want to set the tts to be a different one. By passing
those objects at Run() call time, they are not bounded in any way and
enables greater flexibility.