Moves the speech rate configuration inside the TTS class and adds helper methods.
Moves the speech rate configuration from the server state inside the TTS
class. this makes more sense logically because the speech rate is an
attribute of the tts, and not part of the server state. this is also one
of the changes that I intend to make to change the appearance of the
server state, which was becoming like a singleton class being passed
everywhere. Also adds a distinction between Say() and Output() (please
see comments of those methods). the clear advantage is that now we can
compute the prefix string only using the TTS class.