Implement an audio task manager.
This will allow the player object to be used for different purposes during
the execution of the program.  It manages a number of audio task objects,
which are enqueued and run in sequence.  When a task object is running, the
main loop will call it continuously until it finishes its playback, which
then releases itself and continues execution of the next task.

The TTS class now simply generates speech synthesis tasks that are sent to
the task manager.  Other kinds of tasks in the future will include tone and
wav file playback.