Adds initial support for text formatting in different levels of punctuation.
Implements the TextFormatter class, which formats the text to be sent to
the user. It has already two modes implemented:
* ALL
* SOME
* TODO: NONE

It is still necessary to test what punctuation will be spoken in the
punctuation mode = NONE and punctuation mode = SOME.

Also implements the command to change the punctuation level inside the
speech server.