Handle small sounds at the beginning of the stream.
Fixes a bug in which if a small task runs and writes a very small
amount of data to the player (i.e. a short tone or a short silence),
the stream would start and immediately underrun before the next task
could have a change to fill in more data.

This commit changes the start threshold of the PCM stream to one full
period. If the stream doesn't start by the end of the enqueued tasks,
force it to start manually.