Fixes the prefix string issue in TTS.
this change moves the logic to deal with the default voice annotation
from outside the PrefixString method. Now, users of the class can select
which voice parameter to use. right now only adult male is supported,
but we can easily add more as defined in the ECI IBM TTS documentation.

Also, fixes the d command, where different voice annotations were being
resetted due to applying the voice annotation after the annotations
added with the c command in the queue.